+++ src/main/kotlin/dev/turingcomplete/intellijbytecodeplugin/bytecode/ClassVersionUtils.kt
@@ -2,6 +2,7 @@ package dev.turingcomplete.intellijbytecodeplugin.bytecode
 
 import dev.turingcomplete.intellijbytecodeplugin.org.objectweb.asm.Opcodes
 
+@Suppress("MemberVisibilityCanBePrivate")
 object ClassVersionUtils {
   // -- Properties -------------------------------------------------------------------------------------------------- //